update package for state module
