+++ module/parser-expressions/src/commonMain/kotlin/org/cru/godtools/shared/tool/parser/expressions/grammar/StateExpressionEvaluator.kt
@@ -3,7 +3,7 @@ package org.cru.godtools.shared.tool.parser.expressions.grammar
 import org.cru.godtools.shared.tool.parser.expressions.grammar.generated.StateExpressionBaseVisitor
 import org.cru.godtools.shared.tool.parser.expressions.grammar.generated.StateExpressionParser
 import org.cru.godtools.shared.tool.parser.expressions.grammar.generated.StateExpressionParser.Tokens
-import org.cru.godtools.tool.state.State
+import org.cru.godtools.shared.tool.state.State
 
 internal class StateExpressionEvaluator(private val state: State) {
     val booleanExpr = object : StateExpressionBaseVisitor<Boolean>() {
