+++ module/parser/src/iosMain/kotlin/org/cru/godtools/tool/model/IosFlow.kt
@@ -1,7 +1,7 @@
 package org.cru.godtools.tool.model
 
 import kotlinx.coroutines.flow.combine
-import org.cru.godtools.tool.state.State
+import org.cru.godtools.shared.tool.state.State
 import org.cru.godtools.tool.util.FlowWatcher
 import org.cru.godtools.tool.util.FlowWatcher.Companion.watch
