+++ module/parser/src/iosMain/kotlin/org/cru/godtools/tool/model/IosMultiselect.kt
@@ -1,6 +1,6 @@
 package org.cru.godtools.tool.model
 
-import org.cru.godtools.tool.state.State
+import org.cru.godtools.shared.tool.state.State
 import org.cru.godtools.tool.util.FlowWatcher.Companion.watch
 
 fun Multiselect.Option.watchIsSelected(state: State, block: (Boolean) -> Unit) = isSelectedFlow(state).watch(block)
