--- module/state/src/androidMain/java/org/cru/godtools/tool/internal/Parcelable.kt
@@ -1,4 +1,0 @@
-package org.cru.godtools.tool.internal
-
-actual typealias Parcelable = android.os.Parcelable
-actual typealias Parcelize = kotlinx.parcelize.Parcelize
