+++ module/state/src/androidMain/kotlin/org/cru/godtools/shared/tool/state/internal/Parcelable.kt
@@ -1,0 +1,4 @@
+package org.cru.godtools.shared.tool.state.internal
+
+actual typealias Parcelable = android.os.Parcelable
+actual typealias Parcelize = kotlinx.parcelize.Parcelize
