+++ module/state/src/commonMain/kotlin/org/cru/godtools/shared/tool/state/internal/Parcelable.kt
@@ -1,0 +1,9 @@
+package org.cru.godtools.shared.tool.state.internal
+
+expect interface Parcelable
+
+@OptIn(ExperimentalMultiplatform::class)
+@OptionalExpectation
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.BINARY)
+expect annotation class Parcelize()
