--- module/state/src/commonMain/kotlin/org/cru/godtools/tool/internal/Parcelable.kt
@@ -1,9 +1,0 @@
-package org.cru.godtools.tool.internal
-
-expect interface Parcelable
-
-@OptIn(ExperimentalMultiplatform::class)
-@OptionalExpectation
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.BINARY)
-expect annotation class Parcelize()
