--- module/state/src/jsMain/kotlin/org/cru/godtools/tool/internal/Parcelable.kt
@@ -1,3 +1,0 @@
-package org.cru.godtools.tool.internal
-
-actual interface Parcelable
